source-contract embedding: harvestable truth across the engine (SelfDoc step 1)
Accumulated source-embedded contract work: @dottalk.usage / @dottalk.contract
blocks, dotref/foxref-facing reference edits, and structured header/source
comments across ~120 engine files (src/cli, src/edu, src/ext, src/manual,
src/xbase, src/xindex, include/*), plus the accompanying doc/registry/CI updates
(19 md, workflows, rules, manifests). Also folds in the src/cli code-move
deletions (nav_select.cpp, table_state.hpp, table_write.hpp) coupled to the
edits so a fresh clone stays build-consistent.

This is Stage 1 of the documentation pipeline (SelfDoc harvests it into HELP
DATA -> MDO organizes -> manual/website). Passed the pre-push gate (--allow-mass;
0 hard-blocks). Binary fixture regen, vendored-tree cleanup, and untracked new
work are parked for their own deliberate passes.

+++ include/xbase/index_hooks.hpp
@@ -23,7 +23,7 @@ struct Hooks {
     bool (*apply_replace)(DbArea&,
                           const Snapshot&,
                           const Snapshot&,
-                          std::uint32_t) = nullptr;
+                          std::uint64_t) = nullptr;
     void (*detach)(DbArea&) noexcept = nullptr;
 };
 
@@ -32,7 +32,7 @@ Snapshot capture(DbArea& area);
 bool apply_replace(DbArea& area,
                    const Snapshot& before,
                    const Snapshot& after,
-                   std::uint32_t recno);
+                   std::uint64_t recno);
 void detach(DbArea& area) noexcept;
 
 } // namespace index_hooks
